third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-row-inline-end-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 3
  assert_equals: \n<item data-expected-margin-right="0" data-offset-x="148"></item>\nmargin-right expected "0" but got "20"
Harness: the test ran to completion.
